Add numbers to current input and display to screen
The method that I was using to select number buttons on the calculator
was overly complicated. This commit simplifies the approach, removing
the “id” field from each button tag and wrapping the zero button in a
div so I can select it like the others.

This commit also adds methods to:
- Update the number being edited and display it to the screen when a
  number button is pressed
- Format very big and very small numbers as scientific notation without
  editing the underlying data
- Truncate medium-sized numbers to fit on the calculator screen