Add digit input, operator, and equals sign logic
In general, this commit adds:
- The ability to clear the calculator
- Logic to handle digit input
- Logic to handle operator (+, -, x, /) input
- Logic to handle equals sign (=) input

Previously, there was no way to render a blank display if there was no
input. This commit adds this functionality.

Most of the functions were missing docstrings. This commit adds them.

Number types were previously limited to integers. This commit adds float
arithmetic.

There was no error handling for "divide-by-zero" scenarios. This commit
adds it. It also spruces up error handling by throwing errors in the
low-level functions and handling them in the event listener functions.

Bug fix: the ID of clear button was overriding the built-in `clear()`
function. This commit renames the ID to deconflict.